add docker build for server

+++ scripts/package.mjs
@@ -8,8 +8,14 @@ if (Fs.existsSync(`${pkg}/publish`) && Fs.statSync(`${pkg}/publish`).isDirectory
 Fs.mkdirSync(`${pkg}/publish`);
 Fs.cpSync(`${pkg}/dist`, `${pkg}/publish/dist`, { recursive: true });
 Fs.cpSync(`${pkg}/src`, `${pkg}/publish/src`, { recursive: true });
-Fs.cpSync(`${pkg}/README.md`, `${pkg}/publish/README.md`);
-Fs.cpSync(`${pkg}/LICENSE`, `${pkg}/publish/LICENSE`);
+
+if (Fs.existsSync(`${pkg}/README.md`)) {
+  Fs.cpSync(`${pkg}/README.md`, `${pkg}/publish/README.md`);
+}
+
+if (Fs.existsSync(`${pkg}/LICENSE`)) {
+  Fs.cpSync(`${pkg}/LICENSE`, `${pkg}/publish/LICENSE`);
+}
 
 // TODO: Generate this a bit smarter.
 const pkgJson = JSON.parse(Fs.readFileSync(`${pkg}/package.json`, 'utf-8'));
@@ -22,6 +28,7 @@ const publishPkgJson = {
   url: pkgJson.url,
   directory: pkgJson.directory,
   license: pkgJson.license,
+  main: pkgJson.main,
   module: pkgJson.module,
   types: pkgJson.types,
   sideEffects: pkgJson.sideEffects,
